fix the mirror
I think it works
