+++ Final Project/Assets/Materials/Mirror1-1.renderTexture.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b503eaec7ee9434479b84b4e46eb2365
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 8400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
