--- Final Project/Assets/Materials/Mirror1.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0af1471e20f07464ca87616e8fe96a92
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
