--- Final Project/Assets/Script/MirroMovement.cs
@@ -1,25 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class MirroMovement : MonoBehaviour
-{
-    public Transform playerTarget;
-    public Transform mirror;
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        Vector3 localPlayer = mirror.InverseTransformPoint(playerTarget.position);
-        transform.position = mirror.TransformPoint(new Vector3(localPlayer.x, localPlayer.y - 10, -localPlayer.z));
-
-        // Compute mirrored look-at target
-        Vector3 lookatmirror = mirror.TransformPoint(new Vector3(-localPlayer.x, localPlayer.y - 10, localPlayer.z));
-        transform.LookAt(lookatmirror);
-    }
-}
