--- Final Project/Assets/Script/MirroMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f5db2430909f86e409c6eca5dcb4fa18
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
