+++ Final Project/ProjectSettings/TagManager.asset
@@ -13,9 +13,9 @@ TagManager:
   - 
   - Water
   - UI
-  - MirrorUnseen
-  - 
-  - 
+  - MirrorUnseen2
+  - MirrorUnseen1
+  - MirrorUnseen3
   - 
   - 
   -
